[LoopInterchange] Don't support loops having number of load-store instructions >16
As of now, there is a check which would kick in
if size of DepMatrix is greater than MaxInstrCount.
Size of DepMatrix is dependent on number of load-store instructions,
and levels from `Dependence`. In case where load-store instructions
are too many, size of the matrix would be bigger than MaxInstrCount but
we still spend time in populating the matrix before terminating eventually.

Computing the size of DepMatrix upfront is not possible
in all cases so we turn turn the check to just load-store
instructions. In this patch, we bail out if the loop has more than
16 load-store instructions.